model-server/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

torch>=2.3.0
transformers>=4.44.0
Pillow>=10.0.0
accelerate>=0.33.0
qwen-vl-utils>=0.0.8

pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0

loguru>=0.7.0
prometheus-fastapi-instrumentator>=7.0.0
